Use the X display number, not the Screen number, for the Systray
When we set-up the Systray widget, we want to be the
`_NET_SYSTEM_TRAY_S#` atom on the X display that we are on. This reverts
a change that instead used the screen number, which is not what we want
in multimonitor setups, as both screens are on the same display.

Basically reverts 32a2d4cf277d9d22e84f74d9e27f2b6adfe2b3dd, with an
added comment