Big Floating reorganization/multimonitor bug-fix
Re-organize the use  of Floating layouts to be use single layout for
every group, rather than cloning a new Floating layout for each group.
Some helper functions are added and some function calls changed to
select clients within a single group.

Fix the handling of floating window locations so:
- new floating windows (e.g. dialogs or windows triggered by hooks) will
  properly spawn on the screen corresponding to the given group (rather
  than always spawning relative to x=y=0)
- when a group moves from one screen to another, it's floating windows
  will be moved to lie entirely on the screen
- when a floating window is dragged from one screen to another, the
  window is placed in that group
These in particular greatly improve usability in multimonitor systems.
Previously, dialogs and floating windows would be drawn on different
screens than the group they were associated with, could go off the
screen when moving from a large to a small screen, and could not be
dragged from one monitor to the other (and would cause a nasty snag if
you tried). This should fix all these problems: